@@ -10,7 +10,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>org.pf4j</groupId>
     <artifactId>pf4j-parent</artifactId>
-    <version>2.7.0-SNAPSHOT</version>
+    <version>3.0.0-SNAPSHOT</version>
     <packaging>pom</packaging>
     <name>PF4J Parent</name>
     <description>Plugin Framework for Java</description>
@@ -43,11 +43,12 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>1.7</java.version>
+        <maven.compiler.release>8</maven.compiler.release>
+
         <slf4j.version>1.7.25</slf4j.version>
         <asm.version>7.1</asm.version>
 
-        <junit.version>4.12</junit.version>
+        <junit.version>5.4.0</junit.version>
         <hamcrest.version>2.1</hamcrest.version>
         <mockito.version>2.24.0</mockito.version>
         <cobertura.version>2.7</cobertura.version>
@@ -66,8 +67,6 @@
                     <version>3.8.0</version>
                     <configuration>
                         <showWarnings>true</showWarnings>
-                        <source>${java.version}</source>
-                        <target>${java.version}</target>
                     </configuration>
                 </plugin>
 
@@ -161,6 +160,7 @@
     <modules>
         <module>pf4j</module>
         <module>demo</module>
+        <module>maven-archetypes/quickstart</module>
     </modules>
 
     <profiles>
