@@ -48,14 +48,7 @@ public class BasePluginRepository implements PluginRepository {
         this.filter = filter;
 
         // last modified file is first
-        this.comparator = new Comparator<File>() {
-
-            @Override
-            public int compare(File o1, File o2) {
-                return (int) (o2.lastModified() - o1.lastModified());
-            }
-
-        };
+        this.comparator = (o1, o2) -> (int) (o2.lastModified() - o1.lastModified());
     }
 
     public void setFilter(FileFilter filter) {
@@ -94,13 +87,17 @@ public class BasePluginRepository implements PluginRepository {
 
     @Override
     public boolean deletePluginPath(Path pluginPath) {
+        if (!filter.accept(pluginPath.toFile())) {
+            return false;
+        }
+
         try {
             FileUtils.delete(pluginPath);
             return true;
-        } catch (NoSuchFileException nsf) {
-            return false; // Return false on not found to be compatible with previous API
+        } catch (NoSuchFileException e) {
+            return false; // Return false on not found to be compatible with previous API (#135)
         } catch (IOException e) {
-            throw new RuntimeException(e);
+            throw new PluginRuntimeException(e);
         }
     }
 
