@@ -15,28 +15,34 @@
  */
 package org.pf4j;
 
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
+import org.pf4j.plugin.PluginJar;
 import org.pf4j.plugin.PluginZip;
+import org.pf4j.plugin.TestPlugin;
 
-import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
+import java.io.Writer;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Arrays;
-import java.util.List;
+import java.util.LinkedHashMap;
+import java.util.Map;
+import java.util.Properties;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 /**
  * @author Decebal Suiu
  */
 public class CompoundPluginDescriptorFinderTest {
 
-    @Rule
-    public TemporaryFolder pluginsFolder = new TemporaryFolder();
+    @TempDir
+    Path pluginsPath;
 
     @Test
     public void add() {
@@ -49,8 +55,8 @@ public class CompoundPluginDescriptorFinderTest {
 
     @Test
     public void find() throws Exception {
-        Path pluginPath = pluginsFolder.newFolder("test-plugin-1").toPath();
-        Files.write(pluginPath.resolve("plugin.properties"), getPlugin1Properties(), StandardCharsets.UTF_8);
+        Path pluginPath = Files.createDirectories(pluginsPath.resolve("test-plugin-1"));
+        storePropertiesToPath(getPlugin1Properties(), pluginPath);
 
         PluginDescriptorFinder descriptorFinder = new CompoundPluginDescriptorFinder()
             .add(new PropertiesPluginDescriptorFinder());
@@ -64,31 +70,32 @@ public class CompoundPluginDescriptorFinderTest {
     @Test
     public void findInJar() throws Exception {
         PluginDescriptorFinder descriptorFinder = new CompoundPluginDescriptorFinder()
-            .add(new PropertiesPluginDescriptorFinder());
+            .add(new ManifestPluginDescriptorFinder());
 
-        PluginZip pluginJar = new PluginZip.Builder(pluginsFolder.newFile("my-plugin-1.2.3.jar"), "myPlugin")
+        PluginJar pluginJar = new PluginJar.Builder(pluginsPath.resolve("my-plugin-1.2.3.jar"), "myPlugin")
+            .pluginClass(TestPlugin.class.getName())
             .pluginVersion("1.2.3")
             .build();
 
         PluginDescriptor pluginDescriptor = descriptorFinder.find(pluginJar.path());
         assertNotNull(pluginDescriptor);
         assertEquals("myPlugin", pluginJar.pluginId());
+        assertEquals(TestPlugin.class.getName(), pluginJar.pluginClass());
         assertEquals("1.2.3", pluginJar.pluginVersion());
     }
 
-    @Test(expected = PluginException.class)
-    public void testNotFound() throws Exception {
+    @Test
+    public void testNotFound() {
         PluginDescriptorFinder descriptorFinder = new CompoundPluginDescriptorFinder();
-        Path pluginsPath = pluginsFolder.getRoot().toPath();
-        descriptorFinder.find(pluginsPath.resolve("test-plugin-3"));
+        assertThrows(PluginRuntimeException.class, () -> descriptorFinder.find(pluginsPath.resolve("test-plugin-3")));
     }
 
     @Test
     public void testSpaceCharacterInFileName() throws Exception {
-        PluginDescriptorFinder descriptorFinder = new PropertiesPluginDescriptorFinder();
-        File jar = pluginsFolder.newFile("my plugin-1.2.3.jar");
+        PluginDescriptorFinder descriptorFinder = new CompoundPluginDescriptorFinder()
+            .add(new ManifestPluginDescriptorFinder());
 
-        PluginZip pluginJar = new PluginZip.Builder(jar, "myPlugin")
+        PluginJar pluginJar = new PluginJar.Builder(pluginsPath.resolve("my plugin-1.2.3.jar"), "myPlugin")
             .pluginVersion("1.2.3")
             .build();
 
@@ -96,21 +103,24 @@ public class CompoundPluginDescriptorFinderTest {
         assertNotNull(pluginDescriptor);
     }
 
-    private List<String> getPlugin1Properties() {
-        String[] lines = new String[] {
-            "plugin.id=test-plugin-1\n"
-            + "plugin.version=0.0.1\n"
-            + "plugin.description=Test Plugin 1\n"
-            + "plugin.provider=Decebal Suiu\n"
-            + "plugin.class=org.pf4j.plugin.TestPlugin\n"
-            + "plugin.dependencies=test-plugin-2,test-plugin-3@~1.0\n"
-            + "plugin.requires=>=1\n"
-            + "plugin.license=Apache-2.0\n"
-            + "\n"
-            + ""
-        };
+    private Properties getPlugin1Properties() {
+        Map<String, String> map = new LinkedHashMap<>(7);
+        map.put(PropertiesPluginDescriptorFinder.PLUGIN_ID, "test-plugin-1");
+        map.put(PropertiesPluginDescriptorFinder.PLUGIN_CLASS, TestPlugin.class.getName());
+        map.put(PropertiesPluginDescriptorFinder.PLUGIN_VERSION, "0.0.1");
+        map.put(PropertiesPluginDescriptorFinder.PLUGIN_PROVIDER, "Decebal Suiu");
+        map.put(PropertiesPluginDescriptorFinder.PLUGIN_DEPENDENCIES, "test-plugin-2,test-plugin-3@~1.0");
+        map.put(PropertiesPluginDescriptorFinder.PLUGIN_REQUIRES, ">=1");
+        map.put(PropertiesPluginDescriptorFinder.PLUGIN_LICENSE, "Apache-2.0");
 
-        return Arrays.asList(lines);
+        return PluginZip.createProperties(map);
+    }
+
+    private void storePropertiesToPath(Properties properties, Path pluginPath) throws IOException {
+        Path path = pluginPath.resolve(PropertiesPluginDescriptorFinder.DEFAULT_PROPERTIES_FILE_NAME);
+        try (Writer writer = new OutputStreamWriter(new FileOutputStream(path.toFile()), StandardCharsets.UTF_8)) {
+            properties.store(writer, "");
+        }
     }
 
 }
