@@ -15,14 +15,16 @@
  */
 package org.pf4j;
 
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author Decebal Suiu
@@ -31,7 +33,7 @@ public class DependencyResolverTest {
 
     private DependencyResolver resolver;
 
-    @Before
+    @BeforeEach
     public void init() {
         VersionManager versionManager = new DefaultVersionManager();
         resolver = new DependencyResolver(versionManager);
