@@ -37,19 +37,28 @@ public class ManifestPluginDescriptorFinder implements PluginDescriptorFinder {
 
     private static final Logger log = LoggerFactory.getLogger(ManifestPluginDescriptorFinder.class);
 
+    public static final String PLUGIN_ID = "Plugin-Id";
+    public static final String PLUGIN_DESCRIPTION = "Plugin-Description";
+    public static final String PLUGIN_CLASS = "Plugin-Class";
+    public static final String PLUGIN_VERSION = "Plugin-Version";
+    public static final String PLUGIN_PROVIDER = "Plugin-Provider";
+    public static final String PLUGIN_DEPENDENCIES = "Plugin-Dependencies";
+    public static final String PLUGIN_REQUIRES = "Plugin-Requires";
+    public static final String PLUGIN_LICENSE = "Plugin-License";
+
     @Override
     public boolean isApplicable(Path pluginPath) {
         return Files.exists(pluginPath) && (Files.isDirectory(pluginPath) || FileUtils.isJarFile(pluginPath));
     }
 
     @Override
-    public PluginDescriptor find(Path pluginPath) throws PluginException {
+    public PluginDescriptor find(Path pluginPath) {
         Manifest manifest = readManifest(pluginPath);
 
         return createPluginDescriptor(manifest);
     }
 
-    protected Manifest readManifest(Path pluginPath) throws PluginException {
+    protected Manifest readManifest(Path pluginPath) {
         if (FileUtils.isJarFile(pluginPath)) {
             try (JarFile jar = new JarFile(pluginPath.toFile())) {
                 Manifest manifest = jar.getManifest();
@@ -57,28 +66,28 @@ public class ManifestPluginDescriptorFinder implements PluginDescriptorFinder {
                     return manifest;
                 }
             } catch (IOException e) {
-                throw new PluginException(e);
+                throw new PluginRuntimeException(e);
             }
         }
 
         Path manifestPath = getManifestPath(pluginPath);
         if (manifestPath == null) {
-            throw new PluginException("Cannot find the manifest path");
+            throw new PluginRuntimeException("Cannot find the manifest path");
         }
 
         log.debug("Lookup plugin descriptor in '{}'", manifestPath);
         if (Files.notExists(manifestPath)) {
-            throw new PluginException("Cannot find '{}' path", manifestPath);
+            throw new PluginRuntimeException("Cannot find '{}' path", manifestPath);
         }
 
         try (InputStream input = Files.newInputStream(manifestPath)) {
             return new Manifest(input);
         } catch (IOException e) {
-            throw new PluginException(e);
+            throw new PluginRuntimeException(e);
         }
     }
 
-    protected Path getManifestPath(Path pluginPath) throws PluginException {
+    protected Path getManifestPath(Path pluginPath) {
         if (Files.isDirectory(pluginPath)) {
             // legacy (the path is something like "classes/META-INF/MANIFEST.MF")
             return FileUtils.findFile(pluginPath,"MANIFEST.MF");
@@ -92,37 +101,37 @@ public class ManifestPluginDescriptorFinder implements PluginDescriptorFinder {
 
         // TODO validate !!!
         Attributes attributes = manifest.getMainAttributes();
-        String id = attributes.getValue("Plugin-Id");
+        String id = attributes.getValue(PLUGIN_ID);
         pluginDescriptor.setPluginId(id);
 
-        String description = attributes.getValue("Plugin-Description");
+        String description = attributes.getValue(PLUGIN_DESCRIPTION);
         if (StringUtils.isNullOrEmpty(description)) {
             pluginDescriptor.setPluginDescription("");
         } else {
             pluginDescriptor.setPluginDescription(description);
         }
 
-        String clazz = attributes.getValue("Plugin-Class");
+        String clazz = attributes.getValue(PLUGIN_CLASS);
         if (StringUtils.isNotNullOrEmpty(clazz)) {
             pluginDescriptor.setPluginClass(clazz);
         }
 
-        String version = attributes.getValue("Plugin-Version");
+        String version = attributes.getValue(PLUGIN_VERSION);
         if (StringUtils.isNotNullOrEmpty(version)) {
             pluginDescriptor.setPluginVersion(version);
         }
 
-        String provider = attributes.getValue("Plugin-Provider");
+        String provider = attributes.getValue(PLUGIN_PROVIDER);
         pluginDescriptor.setProvider(provider);
-        String dependencies = attributes.getValue("Plugin-Dependencies");
+        String dependencies = attributes.getValue(PLUGIN_DEPENDENCIES);
         pluginDescriptor.setDependencies(dependencies);
 
-        String requires = attributes.getValue("Plugin-Requires");
+        String requires = attributes.getValue(PLUGIN_REQUIRES);
         if (StringUtils.isNotNullOrEmpty(requires)) {
             pluginDescriptor.setRequires(requires);
         }
 
-        pluginDescriptor.setLicense(attributes.getValue("Plugin-License"));
+        pluginDescriptor.setLicense(attributes.getValue(PLUGIN_LICENSE));
 
         return pluginDescriptor;
     }
