@@ -29,12 +29,9 @@ public interface PluginDescriptorFinder {
 
     /**
      * Returns true if this finder is applicable to the given {@link Path}.
-     *
-     * @param pluginPath
-     * @return
      */
     boolean isApplicable(Path pluginPath);
 
-    PluginDescriptor find(Path pluginPath) throws PluginException;
+    PluginDescriptor find(Path pluginPath);
 
 }
