@@ -19,7 +19,7 @@ import java.nio.file.Path;
 import java.util.List;
 
 /**
- * Directory that contains plugins. A plugin could be a zip file.
+ * Directory that contains plugins. A plugin could be a {@code directory}, @code zip} or {@code jar} file.
  *
  * @author Decebal Suiu
  * @author Mário Franco
@@ -29,7 +29,7 @@ public interface PluginRepository {
     /**
      * List all plugin paths.
      *
-     * @return a list of files
+     * @return a list with paths
      */
     List<Path> getPluginPaths();
 
@@ -38,6 +38,7 @@ public interface PluginRepository {
      *
      * @param pluginPath the plugin path
      * @return true if deleted
+     * @throws PluginRuntimeException if something goes wrong
      */
     boolean deletePluginPath(Path pluginPath);
 
