@@ -41,6 +41,7 @@ public class PluginWrapper {
         this.pluginClassLoader = pluginClassLoader;
 
         pluginState = PluginState.CREATED;
+        runtimeMode = pluginManager.getRuntimeMode();
     }
 
     /**
@@ -120,11 +121,9 @@ public class PluginWrapper {
         }
 
         PluginWrapper other = (PluginWrapper) obj;
-        if (!descriptor.getPluginId().equals(other.descriptor.getPluginId())) {
-            return false;
-        }
 
-        return true;
+        return descriptor.getPluginId().equals(other.descriptor.getPluginId());
+
     }
 
     @Override
@@ -134,10 +133,6 @@ public class PluginWrapper {
 
     void setPluginState(PluginState pluginState) {
         this.pluginState = pluginState;
-    }
-
-    void setRuntimeMode(RuntimeMode runtimeMode) {
-        this.runtimeMode = runtimeMode;
     }
 
     void setPluginFactory(PluginFactory pluginFactory) {
