@@ -23,11 +23,9 @@ import java.util.Collections;
 import java.util.List;
 
 /**
- * This filter providing conditional AND logic across a list of
- * file filters. This filter returns <code>true</code> if all filters in the
- * list return <code>true</code>. Otherwise, it returns <code>false</code>.
- * Checking of the file filter list stops when the first filter returns
- * <code>false</code>.
+ * This filter providing conditional AND logic across a list of file filters.
+ * This filter returns {@code true} if all filters in the list return {@code true}. Otherwise, it returns {@code false}.
+ * Checking of the file filter list stops when the first filter returns {@code false}.
  *
  * @author Decebal Suiu
  */
@@ -37,7 +35,7 @@ public class AndFileFilter implements FileFilter {
     private List<FileFilter> fileFilters;
 
     public AndFileFilter() {
-        this(new ArrayList<FileFilter>());
+        this(new ArrayList<>());
     }
 
     public AndFileFilter(FileFilter... fileFilters) {
@@ -68,7 +66,7 @@ public class AndFileFilter implements FileFilter {
 
     @Override
     public boolean accept(File file) {
-        if (this.fileFilters.size() == 0) {
+        if (this.fileFilters.isEmpty()) {
             return false;
         }
 
